move more sugared values to sugared_value.h
